refactoring checkError function
